orders/requirements.txt
fastapi==0.58.0
uvicorn==0.11.5
tortoise-orm==0.16.13

# for test purposes
flake8==3.8.3
ipdb==0.13.2
ipython==7.15.0
